# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation of my RESTful API - Todo App that allows clients to manage tasks (todos) through standard HTTP methods.\nEach todo is treated as a resource, identified by a unique ID, and manipulated using predictable endpoints.\nAuthentication is handled via **JWT tokens**.\n\nTo access protected endpoints, a valid token must be included in the `Authorization` header: `Authorization: Bearer \u003cyour_token\u003e`",
        "title": "TODO-App API Documentation w/ Swagger",
        "contact": {
            "name": "Nada Serag",
            "email": "nadaserag2006@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
        "/category/{category}": {
            "get": {
                "description": "Returns a single todo by its category",
                "tags": [
                    "Todos"
                ],
                "summary": "Get a todo by category",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Category to search for: ",
                        "name": "category",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.Todo"
                            }
                        }
                    },
                    "404": {
                        "description": "Category Invalid",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorJSON"
                        }
                    }
                }
            }
        },
        "/status/{status}": {
            "get": {
                "description": "Returns a single todo by its status",
                "tags": [
                    "Todos"
                ],
                "summary": "Get a todo by status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Status to search for: ",
                        "name": "status",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.Todo"
                            }
                        }
                    },
                    "404": {
                        "description": "No todos with the status",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorJSON"
                        }
                    }
                }
            }
        },
        "/todos": {
            "get": {
                "description": "Returns a list of all todos",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Todos"
                ],
                "summary": "List all todos",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.Todo"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Adding/Creating new todo to out table \"todos\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Todos"
                ],
                "summary": "Adds todos",
                "parameters": [
                    {
                        "description": "Todo to create",
                        "name": "todo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.Todo"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/main.Todo"
                        }
                    },
                    "400": {
                        "description": "Invalid json, or missing fields, or invalid priority/category or past due date",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorJSON"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorJSON"
                        }
                    }
                }
            }
        },
        "/todos/search/": {
            "get": {
                "description": "Returns todos that have name-match with the search string entered",
                "tags": [
                    "Todos"
                ],
                "summary": "Search Todos by title",
                "parameters": [
                    {
                        "type": "string",
                        "description": "search string: ",
                        "name": "q",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.Todo"
                            }
                        }
                    },
                    "404": {
                        "description": "No todos match search",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorJSON"
                        }
                    }
                }
            }
        },
        "/todos/{id}": {
            "get": {
                "description": "Returns a single todo by its ID",
                "tags": [
                    "Todos"
                ],
                "summary": "Get a todo by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Todo ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Todo"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorJSON"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.ErrorJSON": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "main.Todo": {
            "type": "object",
            "required": [
                "category",
                "completed",
                "priority",
                "title"
            ],
            "properties": {
                "category": {
                    "type": "string"
                },
                "completed": {
                    "type": "boolean"
                },
                "completedAt": {
                    "type": "string"
                },
                "dueDate": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "priority": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}